requirements.txt
amqp==5.2.0
argh==0.31.2
asgiref==3.8.1
billiard==4.2.0
black==24.4.2
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==7.5.3
cryptography==42.0.8
defusedxml==0.8.0rc2
Django==5.0.6
django-appconf==1.0.6
django-autoslug==1.9.9
django-celery-email==3.0.0
django-countries==7.6.1
django-environ==0.11.2
django-filter==24.2
django-phonenumber-field==7.3.0
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
factory-boy==3.3.0
Faker==25.8.0
flake8==7.1.0
flower==2.0.1
humanize==4.9.0
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
kombu==5.3.7
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.1
pathspec==0.12.1
phonenumbers==8.13.39
pillow==10.3.0
platformdirs==4.2.2
pluggy==1.5.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pycodestyle==2.12.0
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.8.0
pytest==8.2.2
pytest-cov==5.0.0
pytest-django==4.8.0
pytest-factoryboy==2.7.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.6
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
watchdog==4.0.1
wcwidth==0.2.13
